api_yamdb/requirements.txt
asgiref==3.7.2
Django>=3.0.3
django-filter>=2.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt>=5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
PyJWT>=2.3.0
pytz==2022.1
sqlparse==0.4.4
python-dotenv>=0.19.2
pytest-pythonpath==0.7.3
pytest==6.2.4
pytest-django==4.4.0
drf_yasg==1.21.7